Import data entries from old template into most recent files
Template has changes due to changes in the protocol. This functon allows
use to import data entries from the old template into the new one. Note
that we have to do two imports for KJK and RK, since the import script
that was run previously (only on these two) contained a bug that led to
extra rows (due to multiple matches, and due to the lack of hyperlinks
in the title column in RK's version). These bugs are fixed now, and we
will run this version to create a new file containing (almost) all of
the old and new data entries. (Almost, because in case of conflicts or
duplicates, we do not import anything.)