Transition reference coding template from xlsx to csv format
Importing ods files drops formulas in cells, so that any hyperlinking is
lost. This makes it harder to merge that ods imports with csv imports,
where formulas are just text and hence fully preserved.

Best way forward is to have all templates in csv format and convert
to/from ods manually for data entry. It makes sense to automate ods <->
csv conversion at some point, but at least with Python this is
surprisingly hard (because of the formulas) and does currently not seem
important enough.